feat: reorder folder of the components
