+++ src/ui/organisms/image-viewer/index.module.css
