--- src/ui/organisms/product-detail/index.module.css
@@ -1,15 +1,0 @@
-.productInfoContainer {
-  border: 0.5px solid rgba(0, 0, 0, 0.12);
-  border-radius: 10px;
-  padding: 15px;
-}
-
-@media only screen and (max-width: 900px) {
-  .productInfoContainer{
-    margin-top: 20px !important;
-    border: none;
-    border-top: 0.5px solid rgba(0, 0, 0, 0.12);;
-    border-radius: 0px;
-    padding: 15px;
-  }
-}
